+++ pyproject.toml
@@ -28,4 +28,5 @@ dev = [
     "pytest>=8.0.0",
     "flake8>=7.0.0",
     "black>=24.1.1",
+    "types-psycopg2>=2.9.21.20240417",
 ]
