+++ src/open_source_python_template/db/db.py
@@ -1,7 +1,14 @@
 from psycopg2 import pool
-from db_config import DBConfig
 from datetime import datetime, timezone
 
+
+
+class DBConfig:
+    DATABASE_HOST = ""
+    DATABASE_NAME = "task-db"
+    DATABASE_USER = "postgres"
+    DATABASE_PASSWORD = ""
+    DATABASE_PORT = "5432"
 
 class Database:
