+++ src/open_source_python_template/db/db_config.py
@@ -1,6 +1,1 @@
-class DBConfig:
-    DATABASE_HOST = ""
-    DATABASE_NAME = "task-db"
-    DATABASE_USER = "postgres"
-    DATABASE_PASSWORD = ""
-    DATABASE_PORT = "5432"
+
